estacion-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3